Fix Exception Handling in xDisk
The xDisk DSC Resource does not handle exceptions correctly. It
relied on `$($Error[0])` to display the error, but does not work
as the built in variable has not been populated yet with the exception.
We fixed this by capturing the Message property of the caught Exception
object and displaying that.

Fixes Issue #29